chore: extract type as types.ts
