+++ packages/data-loader/src/parsers/parseCsv/formatToKintoneRecords.ts
@@ -1,7 +1,6 @@
-import { FieldsJson } from "../../printers/printAsCsv";
-import { ParsedRecord, CsvRecords } from "./index";
 import { isImportSupportedFieldType } from "./isImportSupportedFieldType";
 import { formatToRecordValue } from "./formatToRecordValue";
+import { CsvRecords, FieldsJson, ParsedRecord } from "../../types";
 
 export const formatToKintoneRecords = ({
   records,
