+++ packages/data-loader/src/printers/printAsCsv/__tests__/extractFieldValue.test.ts
@@ -1,5 +1,5 @@
 import { extractFieldValue } from "../extractFieldValue";
-import { KintoneRecord } from "../index";
+import { KintoneRecord } from "../../../types";
 
 const subTableRecords: KintoneRecord[] = require("./fixtures/subtable_input.json");
