+++ packages/data-loader/src/printers/printAsCsv/index.ts
@@ -1,13 +1,5 @@
-import {
-  KintoneFormFieldProperty,
-  KintoneRecordField,
-} from "@kintone/rest-api-client";
 import { convertKintoneRecordsToCsv } from "./convertKintoneRecordsToCsv";
-
-export type KintoneRecord = { [fieldCode: string]: KintoneRecordField.OneOf };
-export type FieldsJson = {
-  properties: { [k: string]: KintoneFormFieldProperty.OneOf };
-};
+import { FieldsJson, KintoneRecord } from "../../types";
 
 export const printAsCsv = (
   records: KintoneRecord[],
